pyproject.toml
[project]
name = "tdease-backend"
version = "0.1.0"
description = "TDEase Backend - Mass Spectrometry Workflow Execution Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.32.1",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "websockets>=14.1",
    "pandas>=2.2.3",
    "numpy>=2.2.1",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.1",
    # YAML support
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
